# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Family Carpool API
  version: 0.35.0
  description: >
    Contract-first API for family-carpool. Adult authentication uses email
    one-time codes and Bearer sessions. Family circles (one per adult in v1)
    hold kids, named places, activity feeds (Organizer-managed iCal/webcal
    subscriptions with soft-fail sync and server-side background polling),
    manual events (any member may CRUD one-off schedule items with 1+ kids),
    a unified circle calendar read (feed + manual items in a time window) with
    per-adult leave-from (named place, one-time address, or default),
    per-coverage leave-from on active assignments (any member may set),
    per-adult default leave-from, estimated leave-by
    (OSRM or fallback; never live traffic; calendar list does not call
    Nominatim or OSRM — clients fill leave-by via GET …/calendar/leave-by),
    computed driving-block adjacency on calendar items for the viewing adult
    (confirmed carpool legs + leave-by cache venue/duration; FORCE_MERGE /
    FORCE_SPLIT overrides via PUT/DELETE …/calendar/drive-block-overrides),
    a confirmed-ride multi-stop route read/write (GET/PUT
    …/calendar/{source}/{itemId}/route — resolved leave-from → pickups →
    destination; auto-optimized pickup order on build when 2+ middles;
    optional driving-adult middle-stop reorder; cached OSRM pairwise legs;
    estimate only),
    a confirmed-ride Playlist read and Open-in-Spotify handoff
...
    responsibility
...
servers:
...
paths:
...
  /api/family/circle/calendar:
    get:
      operationId: listCircleCalendar
      tags: [family]
      summary: List unified calendar items (feed + manual) in a time window
      description: >
        Any circle member. Returns feed-imported and manual events whose
        startsAt falls in the half-open interval [from, to), ordered by
        startsAt ascending, then source, then id. Feed items include feedId /
        feedName and kidIds from the feed; manual items omit feed fields.
        Each item is enriched for the **current adult** with leave-from
        (coverage leave-from when that adult is actively covering, else item
        override, else membership default / first located — database and
        geocode_cache only). Active coverage rows include their own
        leave-from / leave-by (cache-only on this list). This list does **not** call Nominatim or OSRM.
        leaveByStatus is UNAVAILABLE when origin or destination text is missing
        (NO_ORIGIN / NO_DESTINATION), or when a one-time origin geocode is
        known to have failed (GEOCODE_FAILED on fill-in / mutation only),
        OK when destination geocode and driving
        duration are already in server caches (computed in-process; still an
        estimate — never live traffic), and PENDING when a live geocode or
        OSRM lookup is still required (including uncached one-time origins).
        Clients must GET
        /api/family/circle/calendar/leave-by to fill PENDING rows. Each item
        also includes coverage assignments for the circle (all members'
        assignments on that item), per-kid rsvps (YES / NO / NO_RESPONSE;
        missing row materializes as NO_RESPONSE), uncoveredKidIds (in-play
        kids — YES or NO_RESPONSE — not covered by any active PENDING or
        CONFIRMED assignment; RSVP NO kids are never uncovered), and conflicts
        (server-computed kid time-overlaps ignore RSVP NO kids; adult coverage
        overlaps using event startsAt/endsAt only; peers may lie outside the
        requested window). When the viewing adult is the confirmed driver on the
        carpool TO leg, items also include driveBlockLinks — time-adjacent
        confirmed-driving TO siblings with combined vs splittable state (auto
        merge rule, or FORCE_MERGE / FORCE_SPLIT override). Interim Agenda copy
        is TO-only (household confirms often set TO+FROM; dual links were
        duplicate noise). Each link includes otherTitle + otherStartsAt (event
        start, not leave-by). Driving blocks are still computed for TO and FROM
        on each read from live ride legs + leave-by cache (no Nominatim/OSRM on
        this list); overrides are the only persisted block state. Empty array
        when none match. from must be strictly before to (400 otherwise). Adults
        without a circle receive 404. Does not mutate feed snapshots or manual
        events.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive window start (UTC instant)
          schema:
            type: string
            format: date-time
          example: "2026-08-01T00:00:00Z"
        - name: to
          in: query
          required: true
          description: Exclusive window end (UTC instant)
          schema:
            type: string
            format: date-time
          example: "2026-09-01T00:00:00Z"
      responses:
        "200":
          description: Calendar items in the window (possibly empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarItem"
              example:
                - id: 01900000-0000-7000-8000-000000000051
                  source: MANUAL
                  title: Dentist
                  startsAt: "2026-08-15T16:00:00Z"
                  endsAt: null
                  location: Clinic
                  kidIds:
                    - 01900000-0000-7000-8000-000000000021
                  feedId: null
                  feedName: null
                  leaveFromPlaceId: 01900000-0000-7000-8000-000000000031
                  leaveFromPlaceName: Mom's house
                  leaveFromAddress: null
                  leaveByAt: null
                  leaveByStatus: PENDING
                  leaveByReason: null
                  coverages:
                    - id: 01900000-0000-7000-8000-000000000071
                      coveringAdultId: 01900000-0000-7000-8000-000000000001
                      coveringAdultDisplayName: Alex
                      assignedByAdultId: 01900000-0000-7000-8000-000000000002
                      kidIds:
                        - 01900000-0000-7000-8000-000000000021
                      status: CONFIRMED
                      leaveFromPlaceId: 01900000-0000-7000-8000-000000000031
                      leaveFromPlaceName: Mom's house
                      leaveFromAddress: null
                      leaveByAt: null
                      leaveByStatus: PENDING
                      leaveByReason: null
                  uncoveredKidIds: []
                  conflicts: []
                  rsvps:
                    - kidId: 01900000-0000-7000-8000-000000000021
                      status: YES
                - id: 01900000-0000-7000-8000-000000000061
                  source: FEED
                  title: Practice
                  startsAt: "2026-08-15T17:00:00Z"
                  endsAt: "2026-08-15T18:00:00Z"
                  location: Field 3
                  kidIds:
                    - 01900000-0000-7000-8000-000000000021
                  feedId: 01900000-0000-7000-8000-000000000041
                  feedName: U12
                  leaveFromPlaceId: null
                  leaveFromPlaceName: null
                  leaveFromAddress: null
                  leaveByAt: null
                  leaveByStatus: UNAVAILABLE
                  leaveByReason: NO_ORIGIN
                  coverages: []
                  uncoveredKidIds:
                    - 01900000-0000-7000-8000-000000000021
                  conflicts:
                    - type: KID_TIME_OVERLAP
                      kidId: 01900000-0000-7000-8000-000000000021
                      otherKidId: null
                      adultId: null
                      adultDisplayName: null
                      otherSource: MANUAL
                      otherItemId: 01900000-0000-7000-8000-000000000051
                      otherTitle: Dentist
                      otherStartsAt: "2026-08-15T16:00:00Z"
                  rsvps:
                    - kidId: 01900000-0000-7000-8000-000000000021
                      status: NO_RESPONSE
        "400":
          description: Missing from/to, or from is not strictly before to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Current adult has no family circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/family/circle/calendar/leave-by:
    get:
      operationId: listCalendarLeaveBy
      tags: [family]
      summary: Fill estimated leave-by for calendar items in a time window
      description: >
        Any circle member. Returns leave-by enrichment for feed-imported and
        manual events whose startsAt falls in the half-open interval [from, to).
        Item leave-from follows coverage → item override → default → first
        located. Each row also includes leave-from / leave-by for active
        coverage assignments on that item. Uses cached destination and one-time
        origin geocodes and driving durations when present; otherwise may call
        Nominatim and OSRM. Duplicate locations and origin+dest pairs in the
        window share one upstream lookup. Successful geocodes and OSRM
        durations are persisted; misses, OSRM failures, and the fallback
        duration are not. Soft-fail one-time origin geocode → UNAVAILABLE with
        GEOCODE_FAILED (does not fail the whole response). leaveByStatus is OK
        or UNAVAILABLE (never PENDING). Leave-by is always an estimate — never
        live traffic. Empty array when none match. from must be strictly before
        to (400 otherwise). Adults without a circle receive 404. Does not
        mutate feed snapshots or manual events.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive window start (UTC instant)
          schema:
            type: string
            format: date-time
          example: "2026-08-01T00:00:00Z"
        - name: to
          in: query
          required: true
          description: Exclusive window end (UTC instant)
          schema:
            type: string
            format: date-time
          example: "2026-09-01T00:00:00Z"
      responses:
        "200":
          description: Leave-by rows in the window (possibly empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarLeaveBy"
              example:
                - id: 01900000-0000-7000-8000-000000000051
                  source: MANUAL
                  leaveFromPlaceId: 01900000-0000-7000-8000-000000000031
                  leaveFromPlaceName: Mom's house
                  leaveFromAddress: null
                  leaveByAt: "2026-08-15T15:25:00Z"
                  leaveByStatus: OK
                  leaveByReason: null
                  coverages:
                    - id: 01900000-0000-7000-8000-000000000071
                      leaveFromPlaceId: 01900000-0000-7000-8000-000000000031
                      leaveFromPlaceName: Mom's house
                      leaveFromAddress: null
                      leaveByAt: "2026-08-15T15:25:00Z"
                      leaveByStatus: OK
                      leaveByReason: null
                - id: 01900000-0000-7000-8000-000000000061
                  source: FEED
                  leaveFromPlaceId: null
                  leaveFromPlaceName: null
                  leaveFromAddress: null
                  leaveByAt: null
                  leaveByStatus: UNAVAILABLE
                  leaveByReason: NO_ORIGIN
                  coverages: []
        "400":
          description: Missing from/to, or from is not strictly before to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Current adult has no family circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/family/circle/calendar/{source}/{itemId}/playlist:
    get:
      operationId: getCalendarPlaylist
      tags: [family]
      summary: Read Playlist riders for a confirmed ride
      description: >
        Same routability gate as getCalendarRoute (403 when the caller may not
        route). Returns one rider tile per attending kid in the car — driving
        adult's CONFIRMED in-play household kids plus ACCEPTED pickup kids for
        that driver. Connected when any adult has a designated Spotify playlist
        for that kid (viewer designation preferred); includes playlist metadata
        and tracks when fetchable. Unconnected riders include an invite contact
        stub (channel + to) for soft-succeed local UI. Does not mutate state.
      security:
        - bearerAuth: []
      parameters:
        - name: source
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/CalendarItemSource"
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Playlist riders for the confirmed ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarPlaylist"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller may not route this item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Current adult has no family circle, or unknown item id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/family/circle/calendar/{source}/{itemId}/playlist/open:
    post:
      operationId: openCalendarPlaylist
      tags: [family]
      summary: Open-in-Spotify handoff URL for a confirmed ride
      description: >
        Same routability gate as getCalendarRoute. Counts connected kid
        playlists on the ride: 0 → 409; exactly 1 → that playlist's existing
        Spotify URL (no create; viewer need not be connected); 2+ → create or
        replace the viewing adult's private "Carpool merge" playlist with the
        current merge order and return its URL (409 when the viewer lacks
        Spotify). Optional body trackUris supplies a remixed order for 2+;
        when omitted the server applies fair round-robin merge of connected
        riders' tracks. Transient merge playlist id is reused per adult.
      security:
        - bearerAuth: []
      parameters:
        - name: source
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/CalendarItemSource"
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenCalendarPlaylistRequest"
      responses:
        "200":
          description: Spotify web URL to open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarPlaylistOpen"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Caller may not route this item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Current adult has no family circle, or unknown item id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            No connected playlists, viewer Spotify required for 2+ merge but
            not connected, or no track URIs available to merge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/family/circle/calendar/drive-block-overrides:
    put:
      operationId: setDriveBlockOverride
      tags: [family]
      summary: Force-merge or force-split a driving-block pair for the viewing adult
      description: >
        Any circle member, for **their own** overrides only (signed-in adult).
        Persists FORCE_MERGE or FORCE_SPLIT for an ordered adjacent pair
        (earlier = left, later = right; server reorders by startsAt when the
        body is inverted). Driving blocks themselves are not stored — each
        calendar read recomputes from live confirmed carpool legs + leave-by
        cache, then applies this adult's overrides. `createdAt` is set (or
        refreshed) on write so dogfood can count override frequency vs auto
        merges. Returns both calendar items fully re-enriched (including
        driveBlockLinks). Unknown item / no circle → 404. Same left and right
        → 400.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDriveBlockOverrideRequest"
            examples:
              forceSplit:
                summary: Split an auto-combined pair
                value:
                  leg: TO
                  leftSource: FEED
                  leftItemId: 01900000-0000-7000-8000-0000000000a1
                  rightSource: FEED
                  rightItemId: 01900000-0000-7000-8000-0000000000a2
                  action: FORCE_SPLIT
              forceMerge:
                summary: Combine a pair the auto rule split
                value:
                  leg: TO
                  leftSource: FEED
                  leftItemId: 01900000-0000-7000-8000-0000000000a1
                  rightSource: FEED
                  rightItemId: 01900000-0000-7000-8000-0000000000a2
                  action: FORCE_MERGE
      responses:
        "200":
          description: Override saved; both items returned with updated driveBlockLinks
          content:
            application/json:
              schema:
                type: array
                minItems: 2
                maxItems: 2
                items:
                  $ref: "#/components/schemas/CalendarItem"
        "400":
          description: Missing fields or left and right refer to the same item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No circle or unknown calendar item for a source/id in the pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: clearDriveBlockOverride
      tags: [family]
      summary: Clear a driving-block override (restore auto merge)
      description: >
        Any circle member, for **their own** overrides only. Deletes the
        FORCE_MERGE / FORCE_SPLIT row for the ordered pair (server reorders by
        startsAt). Idempotent when nothing was stored. Returns both calendar
        items re-enriched so driveBlockLinks reflect the auto rule again.
      security:
        - bearerAuth: []
      parameters:
        - name: leg
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CarpoolLegKind"
        - name: leftSource
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CalendarItemSource"
        - name: leftItemId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: rightSource
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/CalendarItemSource"
        - name: rightItemId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Override cleared (or was absent); both items returned
          content:
            application/json:
              schema:
                type: array
                minItems: 2
                maxItems: 2
                items:
                  $ref: "#/components/schemas/CalendarItem"
        "400":
          description: Missing fields or left and right refer to the same item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No circle or unknown calendar item for a source/id in the pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/family/circle/calendar/{source}/{itemId}/rsvps/{kidId}:
    parameters:
      - name: source
        in: path
        required: true
        description: Whether the item is a manual event or a feed-imported row
        schema:
          $ref: "#/components/schemas/CalendarItemSource"
      - name: itemId
        in: path
        required: true
        description: Manual event id or feed event id
        schema:
          type: string
          format: uuid
      - name: kidId
        in: path
        required: true
        description: Kid on the calendar item whose RSVP is being set
        schema:
          type: string
          format: uuid
    put:
      operationId: setCalendarRsvp
      tags: [family]
      summary: Set per-kid RSVP on a calendar item
      description: >
        Any circle member. Sets attendance for one kid on the item to YES, NO,
        or NO_RESPONSE. Missing rows read as NO_RESPONSE; setting NO_RESPONSE
        deletes any stored row. Setting NO or NO_RESPONSE while the kid is on
        an active (PENDING or CONFIRMED) coverage assignment hard-releases that
        kid from the assignment (deletes the row if empty). Kid not on the
        item returns 400. Unknown item id for the given source returns 404.
        Returns the enriched calendar item (including rsvps and updated
        coverages / uncoveredKidIds).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCalendarRsvpRequest"
            example:
              status: NO
      responses:
        "200":
          description: RSVP updated; calendar item returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarItem"
        "400":
          description: Kid is not on the calendar item, or status is missing/invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or revoked Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No circle or unknown calendar item for source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Opaque session token returned by verifyAuthCode. Send as
        `Authorization: Bearer <accessToken>`.
  schemas:
...
    CalendarItemSource:
      type: string
      description: Whether the calendar row came from a manual event or a synced feed
      enum:
        - MANUAL
        - FEED
...
    CalendarPlaylist:
      type: object
      required:
        - riders
      properties:
        riders:
          type: array
          description: One tile per attending kid in the car
          items:
            $ref: "#/components/schemas/CalendarPlaylistRider"
    OpenCalendarPlaylistRequest:
      type: object
      properties:
        trackUris:
          type: array
          description: >
            Optional remixed Spotify track URI order for 2+ open handoff. When
            omitted or empty, the server applies fair round-robin merge.
          items:
            type: string
          example: [spotify:track:a1, spotify:track:b1]
    CalendarPlaylistOpen:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Spotify web URL to open in the client
          example: https://open.spotify.com/playlist/stub-playlist-a
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Invalid or expired code
...